requirements.txt
torch>=2.5.0
ninja
numpy
cuda-tile[tileiras]
